The trade itself — silver one way, goods the other, inside one lock window (#459, part 2 of 2 of the server half) (#528)
* feat: buying and selling for silver at the smith, the carpenter, the cook and the trader

Implements #459, part 2 of 2 of the server half — the trade itself.

- Player.Trade moves silver one way and goods the other inside one TryLock
  window on one copy of the slot table, committed only on full success
- uint32 totals: uint16 x uint16 overflows, and 65536 arrows at 1 silver
  would otherwise total zero
- Silver is consumed before the goods are inserted, so paying can free the
  slot the purchase lands in
- consume / consumePack split over a shared consumeWithin, bounded to the
  pack, so nothing worn is ever for sale
- session.go routes PayloadTradeRequest; a refusal never ends the session
- Re-addressing an open stall keeps its revision, now driven by a real trade

* fix: pay a trade total wider than a slot without narrowing it to a uint16

The buying branch compared the purse against a uint32 total and then spent
uint16(total), so a purchase over 65,535 silver was settled for the remainder
of the division. heldInPack sums silver across pack slots and reports a uint32
for that reason; the spending rule now takes one too, so the comparison and the
spend are the same width and no narrowing is left on the paying side.

The comment that stood there argued the opposite of what is true — forty slots
of a uint16 count do not fit a uint16, which is why the truncation was
reachable rather than why it was not.

A restored pack is what makes it reachable: restoredSlots copies a stored count
straight from the record and validateStoredSlot bounds it by nothing but the
uint16 it is stored in, so a purse is not held to a stack maximum.

The selling branch keeps its explicit MaxUint16 refusal: insert writes a single
slot's count and genuinely takes a uint16, which is a different bound from what
a pack-wide spend needs.

* docs: say what the truncation actually reached, not what it could have

The comment claimed the goods were delivered in full. They were not: delivery
still has to fit, and no row of today's price table can put goods worth more
than 65,535 silver into a 36-slot pack, so insert refused and the copy was
discarded. What did get through was the wrong refusal — at a total of 65,536
the low bits are zero, consumePack returned false on its count == 0 guard, and
a purse deep enough to pay was told it was short.

An overclaiming comment is the same defect as the one it replaced, one
direction over instead of one direction under.

+++ server/internal/game/craft.go
@@ -248,7 +248,7 @@ func (t *slotTable) craft(r recipe) bool {
 	scratch := *t
 
 	for _, needed := range r.ingredients {
-		if !scratch.consume(needed.item, needed.count) {
+		if !scratch.consume(needed.item, uint32(needed.count)) {
 			return false
 		}
 	}
